+++ django_imgix/templatetags/imgix_tags.py
@@ -1,36 +1,25 @@
 from django import template
 from django.utils.html import escape
 
-from .. import get_imgix_url
+from .. import get_imgix_url, MAIN_SOURCE_KEY
 
 
 register = template.Library()
 
 
 @register.simple_tag
-def get_imgix(image_url, alias=None, wh=None, **kwargs):
+def get_imgix(image_url, alias=None, wh=None, source=MAIN_SOURCE_KEY, **kwargs):
     """
-    Template tag for returning an Imgix image URL.
+    Template tag tat returns an Imgix image URL.
 
-    In `kwargs`, keys that correspond to `django-imgix` settings will override them. They are:
-
-    - `use_https` overrides `IMGIX_HTTPS`
-    - `web_proxy` overrides `IMGIX_WEB_PROXY`
-    - `domains` overrides `IMGIX_DOMAINS`
-    - `sign_key` overrides `IMGIX_SIGN_KEY`
-    - `shard_strategy` overrides `IMGIX_SHARD_STRATEGY`
-    - `aliases` overrides `IMGIX_ALIASES`
-    - `format_detect` overrides `IMGIX_FORMAT_DETECT`
-
-    Others will be passed to the Imgix API. See https://www.imgix.com/docs/reference
-
-    This template tag returns a string that represents the Imgix URL for the image.
-
-    :param image_url: the image URL that we will pass onto Imgix
+    :param image_url: the image URL that will be passed to Imgix
     :param alias: An optional alias name corresponding to a key in IMGIX_ALIASES setting.
     :param wh: An optional string in the format '600x400'.
+    :param source: An optional string that determines the source to use. Defaults to the
+                   main source.
     :param kwargs: optional, arbitrary number of key-value pairs to override settings
                    and pass to the Imgix API.
+                   See https://www.imgix.com/docs/reference
     :returns: An Imgix URL
     """
-    return escape(get_imgix_url(image_url, alias=alias, wh=wh, **kwargs))
+    return escape(get_imgix_url(image_url, alias=alias, wh=wh, source=source, **kwargs))